+++ src/main/java/pl/plajer/villagedefense/handlers/language/LanguageManager.java
@@ -31,6 +31,7 @@ import java.util.Properties;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
+import org.bukkit.configuration.file.FileConfiguration;
 
 import pl.plajer.villagedefense.Main;
 import pl.plajer.villagedefense.utils.MessageUtils;
@@ -47,6 +48,7 @@ public class LanguageManager {
   private static Main plugin;
   private static Locale pluginLocale;
   private static Properties properties = new Properties();
+  private static FileConfiguration languageConfig;
 
   /**
    * Initializes language management system
@@ -60,6 +62,7 @@ public class LanguageManager {
     if (!new File(plugin.getDataFolder() + File.separator + "language.yml").exists()) {
       plugin.saveResource("language.yml", false);
     }
+    languageConfig = ConfigUtils.getConfig(plugin, "language");
     registerLocales();
     setupLocale();
     //we will wait until server is loaded, we won't soft depend those plugins
@@ -205,7 +208,7 @@ public class LanguageManager {
 
   public static List<String> getLanguageList(String path) {
     if (isDefaultLanguageUsed()) {
-      return ConfigUtils.getConfig(plugin, "language").getStringList(path);
+      return languageConfig.getStringList(path);
     } else {
       return Arrays.asList(plugin.getChatManager().colorMessage(path).split(";"));
     }
@@ -213,7 +216,7 @@ public class LanguageManager {
 
   public static String getLanguageMessage(String message) {
     if (isDefaultLanguageUsed()) {
-      return ConfigUtils.getConfig(plugin, "language").getString(message, "ERR_MESSAGE_NOT_FOUND");
+      return languageConfig.getString(message, "ERR_MESSAGE_NOT_FOUND");
     }
     try {
       return properties.getProperty(ChatColor.translateAlternateColorCodes('&', message));